+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Random damage generator.
+- Generator of Robe of Useful Items's items.
 
 ## [1.1.2] - 2022-05-12
 ### Changed